% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/identify.R
\name{identify}
\alias{identify}
\title{Identify the OAI-PMH service for each data provider.}
\usage{
identify(url = "http://oai.datacite.org/oai", ...)
}
\arguments{
\item{url}{OAI-PMH base url}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Identify the OAI-PMH service for each data provider.
}
\examples{
\dontrun{
identify()

# curl options
library("httr")
identify(config = verbose())
}
}

