% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oai_available.R
\name{oai_available}
\alias{oai_available}
\title{Test of OAI-PMH service is available}
\usage{
oai_available(u, ...)
}
\arguments{
\item{u}{base URL to OAI-PMH service}

\item{...}{other arguments passed to \code{\link{id}}}
}
\value{
\code{TRUE} or \code{FALSE} if the service is available.
}
\description{
Silently test if OAI-PMH service is available under the URL provided.
}
\examples{
\dontrun{
url_list <- list(
  archivesic="http://archivesic.ccsd.cnrs.fr/oai/oai.php",
  datacite = "http://oai.datacite.org/oai",

  # No OAI-PMH here
  google = "http://google.com"
)

sapply(url_list, oai_available)
}
}
