% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oai-package.R
\docType{data}
\name{providers}
\alias{providers}
\title{Metadata providers data.frame.}
\value{
A data.frame of three columns:
\itemize{
 \item repo_name - Name of the OAI repository
 \item base_url - Base URL of the OAI repository
 \item oai_identifier - OAI identifier for the OAI repository
}
}
\description{
Metadata providers data.frame.
}
\keyword{datasets}
