% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/id.R
\name{id}
\alias{id}
\title{Identify the OAI-PMH service for each data provider.}
\usage{
id(url, ...)
}
\arguments{
\item{url}{OAI-PMH base url}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Identify the OAI-PMH service for each data provider.
}
\examples{
\dontrun{
# default, datacite
id()
# arxiv
id("http://export.arxiv.org/oai2")

# curl options
library("httr")
id(config = verbose())
}
}

