% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{id_entify}
\alias{id_entify}
\title{Identify the OAI-PMH service for each data provider.}
\usage{
id_entify(url, ...)
}
\arguments{
\item{url}{OAI-PMH base url. Required.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Identify the OAI-PMH service for each data provider.
}
\examples{
\dontrun{
# datacite
id_entify("http://oai.datacite.org/oai")

# arxiv
id_entify("http://export.arxiv.org/oai2")

# GBIF - http://www.gbif.org/
id_entify("http://api.gbif.org/v1/oai-pmh/registry")

# curl options
library("httr")
id_entify("http://export.arxiv.org/oai2", config = verbose())
}
}

