% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list_records.R
\name{list_records}
\alias{list_records}
\title{List records}
\usage{
list_records(url = "http://oai.datacite.org/oai", prefix = "oai_dc",
  from = NULL, until = NULL, set = NULL, token = NULL, ...)
}
\arguments{
\item{url}{OAI-PMH base url}

\item{prefix}{specifies the metadata format that the records will be
returned in. Default: oai_dc}

\item{from}{specifies that records returned must have been created/update/deleted
on or after this date.}

\item{until}{specifies that records returned must have been created/update/deleted
on or before this date.}

\item{set}{specifies the set that returned records must belong to.}

\item{token}{a token previously provided by the server to resume a request
where it last left off.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
List records
}
\examples{
\dontrun{
list_records(from = '2011-05-01T', until = '2011-08-15T')
list_records(from = '2011-06-01T', until = '2011-07-01T')
}
}

